+++ src/components/Card/example/index.js
@@ -6,19 +6,29 @@ const someMethod = () => { return true; };
 export default [
   {
     examples: [
-      <Card title="this is a title">
-        <div>
-          Hello World
-        </div>
+      <Card title="Title">
+        Hello World
       </Card>,
     ],
   },
   {
     examples: [
-      <Card title="this is a title" onClose={ someMethod } testSection="data-track-card-close">
-        <div>
-          Hello World with closer element
-        </div>
+      <Card title="Title" onClose={ someMethod } testSection="data-track-card-close">
+        Hello World with closer element
+      </Card>,
+    ],
+  },
+  {
+    examples: [
+      <Card>
+        Blank card
+      </Card>,
+    ],
+  },
+  {
+    examples: [
+      <Card shadow={ true }>
+        Shadow
       </Card>,
     ],
   },