+++ src/components/Card/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames';
 
 import Button from '../Button';
 import Icon from 'react-oui-icons';
@@ -18,24 +19,34 @@ const Card = ({
   onClose,
   testSection,
   children,
+  shadow = false,
 }) => {
   const closer = (
     <Button style="unstyled" onClick={ onClose } data-test-section={ `${testSection}-close` }>
       <Icon name='close'/>
     </Button>
   );
+  const shadowClassName = 'oui-shadow--light';
+  const classes = classNames({
+    'push--bottom border--all': true,
+    [`${shadowClassName}`]: shadow,
+  });
 
   return (
-    <div className="push--bottom">
-      <div className="flex background--faint soft border--all soft-half--ends">
-        <div className="width--3-4 epsilon line--loose" data-test-section={ `${testSection}-title` }>
-          { title }
+    <div
+      data-oui-component={ true }
+      className={ classes }>
+      { title && (
+        <div className="flex border--bottom background--faint soft soft-half--ends">
+          <h4 className="flex--1" data-test-section={ `${testSection}-title` }>
+            { title }
+          </h4>
+          <div className="">
+            { onClose && closer }
+          </div>
         </div>
-        <div className="width--1-4 text--right push--right push-half--top">
-          { onClose && closer }
-        </div>
-      </div>
-      <div className="soft border--sides border--bottom" data-test-section={ `${testSection}-body` }>
+      ) }
+      <div className="soft" data-test-section={ `${testSection}-body` }>
         { children }
       </div>
     </div>
@@ -47,6 +58,8 @@ Card.propTypes = {
   children: PropTypes.string.isRequired,
   /** method to invoke when a close element is clicked */
   onClose: PropTypes.func,
+  /** Display a subtle shadow around card. */
+  shadow: PropTypes.boolean,
   /** For automated testing only. */
   testSection: PropTypes.string,
   /** What is the card title */