+++ src/components/Select/example/index.js
@@ -1,0 +1,14 @@
+/* eslint-disable react/jsx-key */
+import React from 'react';
+import Select from '../index';
+
+export default [
+  {
+    examples: [
+      <Select name='zoo' id='zoo'>
+        <option value="one">This is option one</option>
+        <option value="two">And this is option two</option>
+      </Select>,
+    ],
+  },
+];