+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Patch] Add shadow prop to Card component.
+- [Patch] Update example for Select.
+- [Patch] Update new components to include `oui-component` attribute for React highlighting extension.
+- [Patch] Bump react-oui-icons to `1.0.2`.
+
 ## 28.0.3 - 2017-10-18
 - [Patch] Adds Card element to support SPA project